move optimizer to a new mod
